expose the helpers thru src dir
