+++ spec/computable/parameterizer/can-be-set.spec.ts
@@ -1,8 +1,7 @@
 import * as ganache from 'ganache-cli'
 import Web3 from 'web3'
 import { EventEmitter, EventLog } from 'web3/types.d'
-import { increaseTime } from '../../helpers'
-import { onData, eventReturnValues } from '../../../src/helpers'
+import { increaseTime, onData, eventReturnValues } from '../../../src/helpers'
 import Erc20 from '../../../src/contracts/erc-20'
 import Parameterizer from '../../../src/contracts/parameterizer'
 import { Addresses, ParameterDefaults } from '../../../src/constants'
