+++ spec/computable/parameterizer/challenge-can-be-resolved.spec.ts
@@ -1,8 +1,7 @@
 import * as ganache from 'ganache-cli'
 import Web3 from 'web3'
 import { Contract } from 'web3/types.d'
-import { increaseTime } from '../../helpers'
-import { onData, eventReturnValues } from '../../../src/helpers'
+import { increaseTime, onData, eventReturnValues } from '../../../src/helpers'
 import { deployDll, deployAttributeStore } from '../../../src/helpers'
 import { ParameterDefaults } from '../../../src/constants'
 import Erc20 from '../../../src/contracts/erc-20'
