+++ spec/computable/registry/claim-reward.spec.ts
@@ -1,7 +1,6 @@
 import * as ganache from 'ganache-cli'
 import Web3 from 'web3'
 import { Contract, Block } from 'web3/types.d'
-import { stringToBytes, increaseTime } from '../../helpers'
 import Erc20 from '../../../src/contracts/erc-20'
 import Voting from '../../../src/contracts/plcr-voting'
 import Parameterizer from '../../../src/contracts/parameterizer'
@@ -12,7 +11,9 @@ import {
   deployAttributeStore,
   maybeParseInt,
   onData,
-  eventReturnValues
+  eventReturnValues,
+  stringToBytes,
+  increaseTime,
 } from '../../../src/helpers'
 
 let web3:Web3,
