+++ spec/computable/registry/deposit.spec.ts
@@ -1,7 +1,7 @@
 import * as ganache from 'ganache-cli'
 import Web3 from 'web3'
 import { Contract } from 'web3/types.d'
-import { stringToBytes, whitelist } from '../../helpers'
+import { whitelist } from '../../helpers'
 import Erc20 from '../../../src/contracts/erc-20'
 import Voting from '../../../src/contracts/plcr-voting'
 import Parameterizer from '../../../src/contracts/parameterizer'
@@ -13,6 +13,7 @@ import {
   deployDll,
   deployAttributeStore,
   maybeParseInt,
+  stringToBytes,
 } from '../../../src/helpers'
 
 let web3:Web3,
