+++ spec/computable/registry/withdraw.spec.ts
@@ -1,7 +1,6 @@
 import * as ganache from 'ganache-cli'
 import Web3 from 'web3'
 import { Contract, Block } from 'web3/types.d'
-import { stringToBytes, increaseTime } from '../../helpers'
 import Erc20 from '../../../src/contracts/erc-20'
 import Voting from '../../../src/contracts/plcr-voting'
 import Parameterizer from '../../../src/contracts/parameterizer'
@@ -12,6 +11,8 @@ import {
   deployAttributeStore,
   maybeParseInt,
   onData,
+  stringToBytes,
+  increaseTime,
 } from '../../../src/helpers'
 
 let web3:Web3,
@@ -100,7 +101,7 @@ describe('Registry: Withdraw', () => {
     const origDeposit = maybeParseInt(listing.unstakedDeposit)
     expect(origDeposit).toBe(ParameterDefaults.MIN_DEPOSIT)
 
-    // Withdraw 
+    // Withdraw
     try {
       const withdrawAmount = ParameterDefaults.MIN_DEPOSIT/2.0
       const tx3 = await registry.withdraw(listBytes, withdrawAmount, { from: applicant })
