+++ spec/computable/voting/plcr-voting.spec.ts
@@ -2,7 +2,6 @@ import * as ganache from 'ganache-cli'
 import Web3 from 'web3'
 import { Contract } from 'web3/types.d'
 import { ParameterDefaults, NAME } from '../../../src/constants'
-import { stringToBytes } from '../../helpers'
 import Erc20 from '../../../src/contracts/erc-20'
 import Voting from '../../../src/contracts/plcr-voting'
 import Parameterizer from '../../../src/contracts/parameterizer'
@@ -12,6 +11,7 @@ import {
   deployDll,
   deployAttributeStore,
   eventReturnValues,
+  stringToBytes,
 } from '../../../src/helpers'
 
 let web3:Web3,
