+++ spec/helpers/index.ts
@@ -1,2 +1,1 @@
-export * from './general'
 export * from './registry'
